Updated with help functions + error fix
