+++ backend/cmd/scans/scans.go
@@ -1,66 +1,18 @@
 package scans
 
 import (
-	"database/sql"
 	"fmt"
 
 	"github.com/bufbuild/connect-go"
 	_ "github.com/lib/pq"
 
+	"github.com/LarsDepuydt/peno-entrepreneurship-3d-oclusion/cmd/help_functions"
 	threedoclusionv1 "github.com/LarsDepuydt/peno-entrepreneurship-3d-oclusion/gen/proto/threedoclusion/v1"
 )
 
-func ConnectToDataBase() (*sql.DB, error) {
-
-	database, error := sql.Open("postgres", "host=host.docker.internal port=5430 user=docker password=docker1 dbname=patient_server sslmode=disable")
-
-	fmt.Println("Connection succesfull")
-
-	if error != nil {
-		return nil, error
-	}
-
-	return database, nil
-}
-
-func getResponseMaker(database *sql.DB, statement string) ([]int64, []string, []string, error) {
-
-	// Execute the statement with the parameter
-	rows, error := database.Query(statement)
-	if error != nil {
-		return nil, nil, nil, error
-	}
-
-	type RowData struct {
-		id   int64
-		scan string
-		date string
-	}
-
-	var (
-		idArray   []int64
-		scanArray []string
-		dateArray []string
-	)
-
-	for rows.Next() {
-		var rowData RowData
-		error = rows.Scan(&rowData.id, &rowData.scan, &rowData.date)
-		if error != nil {
-			panic(error)
-		}
-		idArray = append(idArray, rowData.id)
-		scanArray = append(scanArray, rowData.scan)
-		dateArray = append(dateArray, rowData.date)
-	}
-
-	return idArray, scanArray, dateArray, nil
-
-}
-
 func AddScan(req *connect.Request[threedoclusionv1.AddScanRequest]) (*connect.Response[threedoclusionv1.AddScanResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 	if database == nil || error != nil {
 		return nil, error
 	}
@@ -68,7 +20,7 @@ func AddScan(req *connect.Request[threedoclusionv1.AddScanRequest]) (*connect.Re
 	defer database.Close()
 
 	// Prepare a statement with placeholders for the values
-	statement, error := database.Prepare("INSERT INTO scan (id, scan_file, scan_date) VALUES ($1, $2, $3)")
+	statement, error := database.Prepare("INSERT INTO scan (scan_file, scan_date) VALUES ($1, $2)")
 
 	if error != nil {
 		return nil, error
@@ -77,12 +29,12 @@ func AddScan(req *connect.Request[threedoclusionv1.AddScanRequest]) (*connect.Re
 	defer statement.Close()
 
 	// Perform database modifications, adding a scan
-	_, error = statement.Exec(req.Msg.Id, req.Msg.ScanFile, req.Msg.ScanDate)
+	_, error = statement.Exec(req.Msg.ScanFile, req.Msg.ScanDate)
 	if error != nil {
 		return nil, error
 	}
 
-	responseMessage := fmt.Sprintf("scan with id: %d added with succes;", req.Msg.Id)
+	responseMessage := fmt.Sprintf("scan with id: %s added with succes;", req.Msg.ScanFile)
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.AddScanResponse{
@@ -95,7 +47,7 @@ func AddScan(req *connect.Request[threedoclusionv1.AddScanRequest]) (*connect.Re
 
 func DeleteScan(req *connect.Request[threedoclusionv1.DeleteScanRequest]) (*connect.Response[threedoclusionv1.DeleteScanResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -122,7 +74,7 @@ func DeleteScan(req *connect.Request[threedoclusionv1.DeleteScanRequest]) (*conn
 
 func GetAllScans(req *connect.Request[threedoclusionv1.GetAllScansRequest]) (*connect.Response[threedoclusionv1.GetAllScansResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -131,7 +83,7 @@ func GetAllScans(req *connect.Request[threedoclusionv1.GetAllScansRequest]) (*co
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT * FROM scan;"
 
-	idArray, scanArray, dateArray, error := getResponseMaker(database, statement)
+	idArray, scanArray, dateArray, error := help_functions.GetResponseMakerScan(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -139,7 +91,6 @@ func GetAllScans(req *connect.Request[threedoclusionv1.GetAllScansRequest]) (*co
 	fmt.Println("Got all scans succesfully")
 
 	res := connect.NewResponse(&threedoclusionv1.GetAllScansResponse{
-		Message:   "Returned all scans",
 		IdData:    idArray,
 		ScanData:  scanArray,
 		ScanDates: dateArray,
@@ -150,7 +101,7 @@ func GetAllScans(req *connect.Request[threedoclusionv1.GetAllScansRequest]) (*co
 
 func GetScanByID(req *connect.Request[threedoclusionv1.GetScanByIDRequest]) (*connect.Response[threedoclusionv1.GetScanByIDResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -160,7 +111,7 @@ func GetScanByID(req *connect.Request[threedoclusionv1.GetScanByIDRequest]) (*co
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT * FROM scan WHERE id = $1;"
 
-	idArray, scanArray, dateArray, error := getResponseMaker(database, statement)
+	idArray, scanArray, dateArray, error := help_functions.GetResponseMakerScan(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -169,7 +120,6 @@ func GetScanByID(req *connect.Request[threedoclusionv1.GetScanByIDRequest]) (*co
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.GetScanByIDResponse{
-		Message:  responseMessage,
 		Id:       idArray[0],
 		ScanData: scanArray[0],
 		ScanDate: dateArray[0],
@@ -180,7 +130,7 @@ func GetScanByID(req *connect.Request[threedoclusionv1.GetScanByIDRequest]) (*co
 
 func GetScanByDate(req *connect.Request[threedoclusionv1.GetScanByDateRequest]) (*connect.Response[threedoclusionv1.GetScanByDateResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -189,7 +139,7 @@ func GetScanByDate(req *connect.Request[threedoclusionv1.GetScanByDateRequest]) 
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT * FROM scan WHERE scan_date = $1;"
 
-	idArray, scanArray, dateArray, error := getResponseMaker(database, statement)
+	idArray, scanArray, dateArray, error := help_functions.GetResponseMakerScan(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -198,7 +148,6 @@ func GetScanByDate(req *connect.Request[threedoclusionv1.GetScanByDateRequest]) 
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.GetScanByDateResponse{
-		Message:   responseMessage,
 		IdData:    idArray,
 		ScanData:  scanArray,
 		ScanDates: dateArray,
