+++ backend/cmd/tags/tags.go
@@ -1,72 +1,25 @@
 package tags
 
 import (
-	"database/sql"
 	"fmt"
 
 	"github.com/bufbuild/connect-go"
 	_ "github.com/lib/pq"
 
+	"github.com/LarsDepuydt/peno-entrepreneurship-3d-oclusion/cmd/help_functions"
 	threedoclusionv1 "github.com/LarsDepuydt/peno-entrepreneurship-3d-oclusion/gen/proto/threedoclusion/v1"
 )
 
-// connect to database function
-// always close database with : defer database.Close()
-func ConnectToDataBase() (*sql.DB, error) {
-
-	database, error := sql.Open("postgres", "host=host.docker.internal port=5430 user=docker password=docker1 dbname=patient_server sslmode=disable")
-
-	fmt.Println("Connection succesfull")
-
-	if error != nil {
-		return nil, error
-	}
-
-	return database, nil
-}
-
-func getResponseMaker(database *sql.DB, statement string) ([]int64, []string, error) {
-
-	// Execute the statement with the parameter
-	rows, error := database.Query(statement)
-	if error != nil {
-		return nil, nil, error
-	}
-
-	type RowData struct {
-		id   int64
-		bite string
-	}
-
-	var (
-		idArray   []int64
-		biteArray []string
-	)
-
-	for rows.Next() {
-		var rowData RowData
-		error = rows.Scan(&rowData.id, &rowData.bite)
-		if error != nil {
-			panic(error)
-		}
-		idArray = append(idArray, rowData.id)
-		biteArray = append(biteArray, rowData.bite)
-	}
-
-	return idArray, biteArray, nil
-
-}
-
 func AddTag(req *connect.Request[threedoclusionv1.AddTagRequest]) (*connect.Response[threedoclusionv1.AddTagResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 	if database == nil || error != nil {
 		return nil, error
 	}
 
 	defer database.Close()
 	// Prepare a statement with placeholders for the values
-	statement, error := database.Prepare("INSERT INTO tag (id, bite) VALUES ($1, $2)")
+	statement, error := database.Prepare("INSERT INTO tag (bite) VALUES ($1)")
 
 	if error != nil {
 		return nil, error
@@ -75,12 +28,12 @@ func AddTag(req *connect.Request[threedoclusionv1.AddTagRequest]) (*connect.Resp
 	defer statement.Close()
 
 	// Perform database modifications, adding a tag
-	_, error = statement.Exec(req.Msg.Id, req.Msg.Bite)
+	_, error = statement.Exec(req.Msg.Bite)
 	if error != nil {
 		return nil, error
 	}
 
-	responseMessage := fmt.Sprintf("tag with id: %d added with succes;", req.Msg.Id)
+	responseMessage := fmt.Sprintf("tag with id: %s added with succes;", req.Msg.Bite)
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.AddTagResponse{
@@ -93,7 +46,7 @@ func AddTag(req *connect.Request[threedoclusionv1.AddTagRequest]) (*connect.Resp
 
 func DeleteTag(req *connect.Request[threedoclusionv1.DeleteTagRequest]) (*connect.Response[threedoclusionv1.DeleteTagResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -120,7 +73,7 @@ func DeleteTag(req *connect.Request[threedoclusionv1.DeleteTagRequest]) (*connec
 
 func GetAllTags(req *connect.Request[threedoclusionv1.GetAllTagsRequest]) (*connect.Response[threedoclusionv1.GetAllTagsResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -129,7 +82,7 @@ func GetAllTags(req *connect.Request[threedoclusionv1.GetAllTagsRequest]) (*conn
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT * FROM tag;"
 
-	idArray, biteArray, error := getResponseMaker(database, statement)
+	idArray, biteArray, error := help_functions.GetResponseMakerTag(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -137,7 +90,6 @@ func GetAllTags(req *connect.Request[threedoclusionv1.GetAllTagsRequest]) (*conn
 	fmt.Println("Got all tags succesfully")
 
 	res := connect.NewResponse(&threedoclusionv1.GetAllTagsResponse{
-		Message:  "Returned all tags",
 		IdData:   idArray,
 		BiteData: biteArray,
 	})
@@ -147,7 +99,7 @@ func GetAllTags(req *connect.Request[threedoclusionv1.GetAllTagsRequest]) (*conn
 
 func GetTagByID(req *connect.Request[threedoclusionv1.GetTagByIDRequest]) (*connect.Response[threedoclusionv1.GetTagByIDResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -156,7 +108,7 @@ func GetTagByID(req *connect.Request[threedoclusionv1.GetTagByIDRequest]) (*conn
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT bite FROM tag WHERE id = $1;"
 
-	idArray, biteArray, error := getResponseMaker(database, statement)
+	idArray, biteArray, error := help_functions.GetResponseMakerTag(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -165,9 +117,8 @@ func GetTagByID(req *connect.Request[threedoclusionv1.GetTagByIDRequest]) (*conn
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.GetTagByIDResponse{
-		Message: responseMessage,
-		Id:      idArray[0],
-		Bite:    biteArray[0],
+		Id:   idArray[0],
+		Bite: biteArray[0],
 	})
 
 	return res, nil
@@ -175,7 +126,7 @@ func GetTagByID(req *connect.Request[threedoclusionv1.GetTagByIDRequest]) (*conn
 
 func GetAllTagsByType(req *connect.Request[threedoclusionv1.GetAllTagsByTypeRequest]) (*connect.Response[threedoclusionv1.GetAllTagsByTypeResponse], error) {
 	// Connect to the database
-	database, error := ConnectToDataBase()
+	database, error := help_functions.ConnectToDataBase()
 
 	if database == nil || error != nil {
 		return nil, error
@@ -184,7 +135,7 @@ func GetAllTagsByType(req *connect.Request[threedoclusionv1.GetAllTagsByTypeRequ
 	// Prepare a statement with placeholders for the condition
 	statement := "SELECT * FROM tag WHERE bite = $1;"
 
-	idArray, biteArray, error := getResponseMaker(database, statement)
+	idArray, biteArray, error := help_functions.GetResponseMakerTag(database, statement)
 	if error != nil {
 		panic(error)
 	}
@@ -193,9 +144,8 @@ func GetAllTagsByType(req *connect.Request[threedoclusionv1.GetAllTagsByTypeRequ
 	fmt.Println(responseMessage)
 
 	res := connect.NewResponse(&threedoclusionv1.GetAllTagsByTypeResponse{
-		Message: responseMessage,
-		Bite:    biteArray[0],
-		IdData:  idArray,
+		Bite:   biteArray[0],
+		IdData: idArray,
 	})
 
 	return res, nil
