+++ frontend/src/gen/proto/threedoclusion/v1/service_pb.js
@@ -13,8 +13,7 @@ import { proto3 } from "@bufbuild/protobuf";
 export const AddTagRequest = proto3.makeMessageType(
   "threedoclusion.v1.AddTagRequest",
   () => [
-    { no: 1, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 2, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 1, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ],
 );
 
@@ -64,9 +63,8 @@ export const GetAllTagsRequest = proto3.makeMessageType(
 export const GetAllTagsResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetAllTagsResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
-    { no: 3, name: "bite_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 1, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
+    { no: 2, name: "bite_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
   ],
 );
 
@@ -86,9 +84,8 @@ export const GetTagByIDRequest = proto3.makeMessageType(
 export const GetTagByIDResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetTagByIDResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 3, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 1, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
+    { no: 2, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ],
 );
 
@@ -108,9 +105,8 @@ export const GetAllTagsByTypeRequest = proto3.makeMessageType(
 export const GetAllTagsByTypeResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetAllTagsByTypeResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 3, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
+    { no: 1, name: "bite", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
   ],
 );
 
@@ -122,9 +118,8 @@ export const GetAllTagsByTypeResponse = proto3.makeMessageType(
 export const AddScanRequest = proto3.makeMessageType(
   "threedoclusion.v1.AddScanRequest",
   () => [
-    { no: 1, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 2, name: "scan_file", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 3, name: "scan_date", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 1, name: "scan_file", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 2, name: "scan_date", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ],
 );
 
@@ -174,10 +169,9 @@ export const GetAllScansRequest = proto3.makeMessageType(
 export const GetAllScansResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetAllScansResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
-    { no: 3, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
-    { no: 4, name: "scan_dates", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 1, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
+    { no: 2, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 3, name: "scan_dates", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
   ],
 );
 
@@ -197,10 +191,9 @@ export const GetScanByIDRequest = proto3.makeMessageType(
 export const GetScanByIDResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetScanByIDResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 3, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 4, name: "scan_date", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 1, name: "id", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
+    { no: 2, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 3, name: "scan_date", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ],
 );
 
@@ -220,10 +213,9 @@ export const GetScanByDateRequest = proto3.makeMessageType(
 export const GetScanByDateResponse = proto3.makeMessageType(
   "threedoclusion.v1.GetScanByDateResponse",
   () => [
-    { no: 1, name: "message", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
-    { no: 3, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
-    { no: 4, name: "scan_dates", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 1, name: "id_data", kind: "scalar", T: 3 /* ScalarType.INT64 */, repeated: true },
+    { no: 2, name: "scan_data", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+    { no: 3, name: "scan_dates", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
   ],
 );
